+++ core/auth/interfaces.go
@@ -18,18 +18,37 @@ type authType interface {
 	//	credentialValue (map): Credential value
 	signUp(authType model.AuthType, appType model.ApplicationType, appOrg model.ApplicationOrganization, creds string, params string, newCredentialID string, l *logs.Log) (string, map[string]interface{}, error)
 
-	//checks the verification code generated on email signup
+	//verifies credential (checks the verification code generated on email signup for email auth type)
 	// Returns:
 	//	authTypeCreds (map[string]interface{}): Updated Credential.Value
-	verify(credential *model.Credential, verification string, l *logs.Log) (map[string]interface{}, error)
+	verifyCredential(credential *model.Credential, verification string, l *logs.Log) (map[string]interface{}, error)
+
+	//restarts the credential verification
+	restartCredentialVerification(credential *model.Credential, l *logs.Log) error
+
 	//sends the verification code to the identifier
 	// Returns:
 	//	authTypeCreds (map[string]interface{}): Updated Credential.Value
 	sendVerify(authType model.AuthType, identifier string, credential *model.Credential, l *logs.Log) (map[string]interface{}, error)
+
+	//updates the value of the credential object with new value
+	// Returns:
+	//	authTypeCreds (map[string]interface{}): Updated Credential.Value
+	resetCredential(credential *model.Credential, resetCode *string, params string, l *logs.Log) (map[string]interface{}, error)
+
+	//apply forgot credential for the auth type (generates a reset password link with code and expiry and sends it to given identifier for email auth type)
+	forgotCredential(credential *model.Credential, identifier string, l *logs.Log) (map[string]interface{}, error)
+
 	//getUserIdentifier parses the credentials and returns the user identifier
 	// Returns:
 	//	userIdentifier (string): User identifier
 	getUserIdentifier(creds string) (string, error)
+
+	//isCredentialVerified says if the credential is verified
+	// Returns:
+	//	verified (bool): is credential verified
+	//	expired (bool): is credential verification expired
+	isCredentialVerified(credential *model.Credential, l *logs.Log) (*bool, *bool, error)
 
 	//checkCredentials checks if the account credentials are valid for the account auth type
 	checkCredentials(accountAuthType model.AccountAuthType, creds string, l *logs.Log) (string, error)
@@ -113,11 +132,41 @@ type APIs interface {
 	//			Params (interface{}): authType-specific set of parameters passed back to client
 	Refresh(refreshToken string, apiKey string, l *logs.Log) (*model.LoginSession, error)
 
-	//Verify checks the verification code in the credentials collection
-	Verify(id string, verification string, l *logs.Log) error
+	//VerifyCredential verifies credential (checks the verification code in the credentials collection)
+	VerifyCredential(id string, verification string, l *logs.Log) error
 
 	//Send verification code to identifier
 	SendVerify(authenticationType string, appTypeIdentifier string, orgID string, apiKey string, identifier string, l *logs.Log) error
+
+	//UpdateCredential updates the credential object with the new value
+	//	Input:
+	//		accountID: id of the associated account to reset
+	//		accountAuthTypeID (string): id of the AccountAuthType
+	//		params: specific params for the different auth types
+	//	Returns:
+	//		error: if any
+	UpdateCredential(accountID string, accountAuthTypeID string, params string, l *logs.Log) error
+
+	//ForgotCredential initiate forgot credential process (generates a reset link and sends to the given identifier for email auth type)
+	//	Input:
+	//		authenticationType (string): Name of the authentication method for provided creds (eg. "email", "username", "illinois_oidc")
+	//		identifier: identifier of the account auth type
+	//		appTypeIdentifier (string): Identifier of the app type/client that the user is logging in from
+	//		orgID (string): ID of the organization that the user is logging in
+	//		apiKey (string): API key to validate the specified app
+	//	Returns:
+	//		error: if any
+	ForgotCredential(authenticationType string, appTypeIdentifier string, orgID string, apiKey string, identifier string, l *logs.Log) error
+
+	//ResetForgotCredential resets forgot credential
+	//	Input:
+	//		credsID: id of the credential object
+	//		resetCode: code from the reset link
+	//		params: specific params for the different auth types
+	//	Returns:
+	//		error: if any
+	ResetForgotCredential(credsID string, resetCode string, params string, l *logs.Log) error
+
 	//GetLoginURL returns a pre-formatted login url for SSO providers
 	//	Input:
 	//		authType (string): Name of the authentication method for provided creds (eg. "email", "username", "illinois_oidc")
@@ -198,6 +247,7 @@ type Storage interface {
 	DeleteExpiredSessions(now *time.Time) error
 
 	//Accounts
+	FindAccountByID(context storage.TransactionContext, storageid string) (*model.Account, error)
 	FindAccount(appOrgID string, authTypeID string, accountAuthTypeIdentifier string) (*model.Account, error)
 	FindAccountByAuthTypeID(context storage.TransactionContext, id string) (*model.Account, error)
 	InsertAccount(account model.Account) (*model.Account, error)
@@ -207,9 +257,10 @@ type Storage interface {
 	FindOrganization(id string) (*model.Organization, error)
 
 	//Credentials
+	InsertCredential(creds *model.Credential) error
 	FindCredential(ID string) (*model.Credential, error)
 	UpdateCredential(creds *model.Credential) error
-	InsertCredential(creds *model.Credential) error
+	UpdateCredentialValue(ID string, value map[string]interface{}) error
 
 	//ServiceRegs
 	FindServiceRegs(serviceIDs []string) ([]model.ServiceReg, error)