+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add password reset from client and reset links [#216](https://github.com/rokwire/core-building-block/issues/216)
+- User PII in tokens [#169](https://github.com/rokwire/core-building-block/issues/169)
 - Add authentication required auth wrapper [#250](https://github.com/rokwire/core-building-block/issues/250)
 
 ## [1.12.0] - 2021-11-10